app/requirements.txt
Flask
mysql-connector
flask-wtf
passlib